+++ test/resque_test.rb
@@ -79,9 +79,9 @@ context "Resque" do
     assert Resque::Job.create(:jobs, 'BadJob', 20, '/tmp')
 
     assert_equal 5, Resque.size(:jobs)
-    Resque::Job.destroy(:jobs, 'SomeJob')
+    assert_equal 2, Resque::Job.destroy(:jobs, 'SomeJob')
     assert_equal 3, Resque.size(:jobs)
-    Resque::Job.destroy(:jobs, 'BadJob', 30, '/tmp')
+    assert_equal 1, Resque::Job.destroy(:jobs, 'BadJob', 30, '/tmp')
     assert_equal 2, Resque.size(:jobs)
   end